refactor: alterado os valores do ADC para 16 bits

+++ src/display/display.h
@@ -5,7 +5,7 @@
 extern ssd1306_t ssd_bm;
 
 extern void configurar_display();
-extern void exibir_leitura_mic(int frequencia_lida, int frequencia_desejada);
+extern void exibir_leitura_mic(float frequencia_lida, float frequencia_desejada);
 extern void exibir_bitmap_display(uint8_t text[]);
 extern void gerenciar_afinacao();
 extern void iniciar_display();
